requirements.txt
asgiref==3.8.1
Django==5.1.4
django-ckeditor==6.7.2
django-cors-headers==4.6.0
django-cuser==2017.3.16
django-jazzmin==3.0.1
django-js-asset==2.2.0
django-modeltranslation==0.19.11
django-rest-framework==0.1.0
django-tinymce==4.1.0
djangorestframework==3.15.2
drf-yasg==1.21.8
inflection==0.5.1
modeltranslation==0.25
packaging==24.2
pillow==11.0.0
pytz==2024.2
PyYAML==6.0.2
sqlparse==0.5.3
tzdata==2024.2
Unidecode==1.3.8
uritemplate==4.1.1